tests/unit_tests/test_api.py
import pytest
from fastapi.testclient import TestClient
from src.agent.api import app

client = TestClient(app)

def test_read_root():
    response = client.get("/")
    assert response.status_code == 200
    assert response.json() == {"message": "Welcome to the Conversational Agent API"}

def test_health_check():
    response = client.get("/health")
    assert response.status_code == 200
    assert response.json() == {"status": "healthy"}

# Add more tests for other API endpoints as needed
