pyproject.toml
[project]
name = "agent"
version = "4.0.0"
description = "Add your description here"
authors = [
    { name = "Marc Fabian Mezger", email = "marc.mezger@gmail.com" }
]

dependencies = [
    "langchain-cohere>=0.3.0",
    "loguru>=0.7.2",
    "omegaconf>=2.3.0",
    "tiktoken>=0.7.0",
    "python-dotenv>=1.0.1",
    "fastapi>=0.115.0",
    "openai>=1.46.0",
    "langchain-qdrant>=0.1.4",
    "uvicorn>=0.30.6",
    "pypdfium2>=4.30.0",
    "ultra-simple-config>=0.1.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.22",
    "arize-phoenix>=4.35.1",
    "fastembed>=0.3.6",
    "nltk>=3.9.1",
    "lingua-language-detector>=2.0.2",
    "weave>=0.51.17",
]

readme = "README.md"
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.3.2",
    "yappi>=1.6.0",
    "mypy>=1.11.2",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/agent"]
